backend/requirements.txt
fastapi
uvicorn-standard
python-multipart
grpcio
grpcio-tools
python-dotenv